Backgrounds implemented, some camera changes
And some beautiful art was added.
Camera changed to hold current active camera, and only the current active camera's matrix is update once per update rather than after every function call which also allows for properly having multiple cameras

+++ game/src/graphics/Particles.cpp
@@ -172,7 +172,7 @@ void ParticleSystem::RenderParticles(glm::vec2 pos)
 	Shaders::particleRenderShader->SetVariable("EmitterPosition", pos);
 	if (m_settings.Texture)
 	{
-		Shaders::particleRenderShader->SetVariable("TextureLayer", m_settings.Texture->GetID());
+		Shaders::particleRenderShader->SetVariable("TextureLayer", static_cast<float>(m_settings.Texture->GetID()) );
 		Shaders::particleRenderShader->SetVariable("TextureBox", glm::vec4(0,0,m_settings.Texture->GetBounds()));
 	}
 	else
